moved usecasegroups in individual files
