+++ webofneeds/won-owner-webapp/src/main/webapp/config/usecase-definitions.js
@@ -1,39 +1,13 @@
-import {
-  isValidNumber,
-  get,
-  getInFromJsonLd,
-  getIn,
-  isValidDate,
-} from "../app/utils.js";
-import { details, abstractDetails } from "detailDefinitions";
-import {
-  realEstateRentRangeDetail,
-  realEstateRentDetail,
-  realEstateFloorSizeDetail,
-  realEstateNumberOfRoomsDetail,
-  realEstateFeaturesDetail,
-  realEstateFloorSizeRangeDetail,
-  realEstateNumberOfRoomsRangeDetail,
-} from "realEstateDetails";
-import { genresDetail, instrumentsDetail } from "musicianDetails";
-import { interestsDetail, skillsDetail } from "personDetails";
-import { findLatestIntervallEndInJsonLdOrNowAndAddMillis } from "../app/won-utils.js";
-import won from "../app/won-es6.js";
-import {
-  filterInVicinity,
-  filterFloorSizeRange,
-  filterNumOfRoomsRange,
-  filterRentRange,
-  concatenateFilters,
-  filterAboutTime,
-  sparqlQuery,
-} from "../app/sparql-builder-utils.js";
+import { details, emptyDraft } from "detailDefinitions";
+import { realEstateUseCases } from "realEstateUseCases";
+import { transportUseCases } from "transportUseCases";
+import { complainUseCases } from "complainUseCases";
+import { socialUseCases } from "socialUseCases";
+import { professionalUseCases } from "professionalUseCases";
+import { mobilityUseCases } from "mobilityUseCases";
+import { musicianUseCases } from "musicianUseCases";
 
-export const emptyDraft = {
-  is: {},
-  seeks: {},
-  matchingContext: undefined,
-};
+import { findLatestIntervallEndInJsonLdOrNowAndAddMillis } from "../app/won-utils.js";
 
 /**
  * USE CASE REQUIREMENTS
@@ -86,1082 +60,6 @@ const allDetailsUseCase = {
     draft: { ...emptyDraft },
     isDetails: details,
     seeksDetails: details,
-  },
-};
-
-// TODO: roles?
-// note: if no details are to be added for is or seeks,
-// there won't be an is or seeks part unless defined in the draft
-// details predefined in the draft can only be changed IF included in the correct detail list
-const socialUseCases = {
-  breakfast: {
-    identifier: "breakfast",
-    label: "Get breakfast",
-    icon: "#ico36_uc_breakfast",
-    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
-    draft: {
-      ...emptyDraft,
-      is: {
-        title: "I'm up for breakfast! Any plans?",
-        tags: ["breakfast"],
-      },
-      seeks: { title: "breakfast" },
-      searchString: "breakfast",
-    },
-    isDetails: {
-      title: { ...details.title },
-      description: { ...details.description },
-      fromDatetime: { ...details.fromDatetime },
-      throughDatetime: { ...details.throughDatetime },
-      location: { ...details.location },
-      interests: { ...interestsDetail },
-    },
-    seeksDetails: undefined,
-  },
-  lunch: {
-    identifier: "lunch",
-    label: "Get lunch",
-    icon: "#ico36_uc_meal-half",
-    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
-    draft: {
-      ...emptyDraft,
-      is: {
-        title: "I'm up for lunch! Any plans?",
-        tags: ["lunch"],
-      },
-      searchString: "lunch",
-    },
-    isDetails: {
-      title: { ...details.title },
-      description: { ...details.description },
-      fromDatetime: { ...details.fromDatetime },
-      throughDatetime: { ...details.throughDatetime },
-      location: { ...details.location },
-      interests: { ...interestsDetail },
-    },
-    seeksDetails: undefined,
-  },
-  afterparty: {
-    identifier: "afterparty",
-    label: "Go out",
-    icon: "#ico36_uc_drinks",
-    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
-    draft: {
-      ...emptyDraft,
-      is: {
-        title: "I'm up for partying! Any plans?",
-        tags: ["afterparty"],
-      },
-      searchString: "afterparty",
-    },
-    isDetails: {
-      title: { ...details.title },
-      fromDatetime: { ...details.fromDatetime },
-      throughDatetime: { ...details.throughDatetime },
-      description: { ...details.description },
-      location: { ...details.location },
-      interests: { ...interestsDetail },
-    },
-  },
-  sightseeing: {
-    identifier: "sightseeing",
-    label: "Go sightseeing",
-    icon: "#ico36_uc_sightseeing",
-    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
-    draft: {
-      ...emptyDraft,
-      is: { tags: ["sightseeing"] },
-      searchString: "sightseeing",
-    },
-    isDetails: {
-      title: { ...details.title },
-      description: { ...details.description },
-      fromDatetime: { ...details.fromDatetime },
-      throughDatetime: { ...details.throughDatetime },
-      location: { ...details.location },
-      interests: { ...interestsDetail },
-    },
-  },
-};
-
-const complainUseCases = {
-  complain: {
-    identifier: "complain",
-    label: "Complain about something",
-    icon: "#ico36_uc_wtf",
-    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
-    draft: {
-      ...emptyDraft,
-      is: {
-        title: "WTF?",
-        tags: ["wtf"],
-      },
-      seeks: {},
-      searchString: "wtf",
-    },
-    isDetails: {
-      title: { ...details.title },
-      description: { ...details.description },
-      location: { ...details.location },
-      tags: { ...details.tags },
-    },
-    seeksDetails: undefined,
-  },
-  handleComplaints: {
-    identifier: "handleComplaints",
-    label: "Handle complaints",
-    icon: "#ico36_uc_wtf_interest",
-    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
-    draft: {
-      ...emptyDraft,
-      is: {
-        title: "I'll discuss complaints",
-      },
-      seeks: {},
-      searchString: "wtf",
-    },
-    isDetails: {
-      title: { ...details.title },
-      description: { ...details.description },
-      location: { ...details.location },
-      tags: { ...details.tags },
-    },
-    seeksDetails: undefined,
-  },
-};
-
-const professionalUseCases = {
-  getToKnow: {
-    identifier: "getToKnow",
-    label: "Find people",
-    icon: "#ico36_uc_find_people",
-    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
-    draft: {
-      ...emptyDraft,
-      is: {
-        title: "I'm up for meeting new people!",
-        tags: ["meetup"],
-      },
-      searchString: "meetup",
-    },
-    isDetails: {
-      title: { ...details.title },
-      description: { ...details.description },
-      location: { ...details.location },
-      person: { ...details.person },
-      skills: { ...skillsDetail },
-      interests: { ...interestsDetail },
-    },
-    seeksDetails: {
-      description: { ...details.description },
-      location: { ...details.location },
-      skills: { ...skillsDetail },
-      interests: { ...interestsDetail },
-    },
-  },
-  phdIs: {
-    identifier: "phdIs",
-    label: "Offer a PhD position",
-    icon: "#ico36_uc_phd",
-    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
-    draft: {
-      ...emptyDraft,
-      is: {
-        title: "I'm offering a PhD position!",
-        tags: ["offer-phd"],
-      },
-      searchString: "search-phd",
-    },
-    isDetails: {
-      title: { ...details.title },
-      description: { ...details.description },
-      location: { ...details.location },
-    },
-    seeksDetails: {
-      skills: { ...skillsDetail, placeholder: "" }, // TODO: find good placeholders
-      interests: { ...interestsDetail, placeholder: "" },
-    },
-  },
-  phdSeeks: {
-    identifier: "phdSeeks",
-    label: "Find a PhD position",
-    icon: "#ico36_uc_phd",
-    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
-    draft: {
-      ...emptyDraft,
-      is: {
-        title: "I'm looking for a PhD position!",
-        tags: ["search-phd"],
-      },
-      searchString: "offer-phd",
-    },
-    isDetails: {
-      title: { ...details.title },
-      description: { ...details.description },
-      person: { ...details.person },
-      skills: { ...skillsDetail, placeholder: "" }, // TODO: find good placeholders
-      interests: { ...interestsDetail, placeholder: "" },
-    },
-    seeksDetails: {
-      description: { ...details.description },
-      location: { ...details.location },
-    },
-  },
-  postDocIs: {
-    identifier: "postDocIs",
-    label: "Offer a PostDoc position",
-    icon: "#ico36_uc_postdoc",
-    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
-    draft: {
-      ...emptyDraft,
-      is: {
-        title: "I'm offering a PostDoc position!",
-        tags: ["offer-postdoc"],
-      },
-      searchString: "search-postdoc",
-    },
-    isDetails: {
-      title: { ...details.title },
-      description: { ...details.description },
-      location: { ...details.location },
-    },
-    seeksDetails: {
-      skills: { ...skillsDetail, placeholder: "" }, // TODO: find good placeholders
-      interests: { ...interestsDetail, placeholder: "" },
-    },
-  },
-  postDocSeeks: {
-    identifier: "postDocSeeks",
-    label: "Find a PostDoc position",
-    icon: "#ico36_uc_postdoc",
-    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
-    draft: {
-      ...emptyDraft,
-      is: {
-        title: "I'm looking for a PostDoc position!",
-        tags: ["search-postdoc"],
-      },
-      searchString: "offer-postdoc",
-    },
-    isDetails: {
-      title: { ...details.title },
-      description: { ...details.description },
-      person: { ...details.person },
-      skills: { ...skillsDetail, placeholder: "" }, // TODO: find good placeholders
-      interests: { ...interestsDetail, placeholder: "" },
-    },
-    seeksDetails: {
-      description: { ...details.description },
-      location: { ...details.location },
-    },
-  },
-  consortiumIs: {
-    identifier: "consortiumIs",
-    label: "Offer slot in a project consortium",
-    icon: "#ico36_uc_consortium-offer",
-    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
-    draft: {
-      ...emptyDraft,
-      is: {
-        title: "Offering a slot in a project consortium",
-        tags: ["offer-consortium"],
-      },
-      searchString: "search-consortium",
-    },
-    isDetails: {
-      title: { ...details.title },
-      description: { ...details.description },
-      location: { ...details.location },
-    },
-    seeksDetails: {
-      description: { ...details.description },
-      location: { ...details.location },
-      skills: { ...skillsDetail, placeholder: "" }, // TODO: find good placeholders
-      interests: { ...interestsDetail, placeholder: "" },
-    },
-  },
-  consortiumSeeks: {
-    identifier: "consortiumSeeks",
-    label: "Find a project consortium to join",
-    icon: "#ico36_uc_consortium-search",
-    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
-    draft: {
-      ...emptyDraft,
-      is: {
-        title: "Looking for a slot in a project consortium",
-        tags: ["search-consortium"],
-      },
-      searchString: "offer-consortium",
-    },
-    isDetails: {
-      title: { ...details.title },
-      description: { ...details.description },
-      location: { ...details.location },
-      skills: { ...skillsDetail, placeholder: "" }, // TODO: find good placeholders
-      interests: { ...interestsDetail, placeholder: "" },
-    },
-    seeksDetails: {
-      description: { ...details.description },
-      location: { ...details.location },
-    },
-  },
-};
-
-const realEstateUseCases = {
-  searchRent: {
-    identifier: "searchRent",
-    label: "Find a place to rent",
-    icon: "#ico36_uc_realestate",
-    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
-    draft: {
-      ...emptyDraft,
-      seeks: {
-        type: "won:RealEstateRentOffer",
-        tags: ["RentOutRealEstate"],
-      },
-      is: {
-        type: "won:RealEstateRentDemand",
-        tags: ["SearchRealEstateToRent"],
-      },
-    },
-    isDetails: undefined,
-    seeksDetails: {
-      location: { ...details.location },
-      floorSizeRange: { ...realEstateFloorSizeRangeDetail },
-      numberOfRoomsRange: { ...realEstateNumberOfRoomsRangeDetail },
-      features: { ...realEstateFeaturesDetail },
-      rentRange: { ...realEstateRentRangeDetail },
-    },
-    generateQuery: (draft, resultName) => {
-      const seeksBranch = draft && draft.seeks;
-      const rentRange = seeksBranch && seeksBranch.rentRange;
-      const floorSizeRange = seeksBranch && seeksBranch.floorSizeRange;
-      const numberOfRoomsRange = seeksBranch && seeksBranch.numberOfRoomsRange;
-      const location = seeksBranch && seeksBranch.location;
-
-      const filters = [
-        {
-          // to select is-branch
-          prefixes: {
-            won: won.defaultContext["won"],
-          },
-          operations: [
-            `${resultName} a won:Need.`,
-            `${resultName} won:is ?is.`,
-            location && "?is won:hasLocation ?location.",
-          ],
-        },
-        rentRange &&
-          filterRentRange(
-            "?is",
-            rentRange.min,
-            rentRange.max,
-            rentRange.currency
-          ),
-
-        floorSizeRange &&
-          filterFloorSizeRange("?is", floorSizeRange.min, floorSizeRange.max),
-
-        numberOfRoomsRange &&
-          filterNumOfRoomsRange(
-            "?is",
-            numberOfRoomsRange.min,
-            numberOfRoomsRange.max
-          ),
-
-        filterInVicinity("?location", location),
-      ];
-
-      const concatenatedFilter = concatenateFilters(filters);
-
-      return sparqlQuery({
-        prefixes: concatenatedFilter.prefixes,
-        selectDistinct: resultName,
-        where: concatenatedFilter.operations,
-        orderBy: [
-          {
-            order: "ASC",
-            variable: "?location_geoDistance",
-          },
-        ],
-      });
-    },
-  },
-  offerRent: {
-    identifier: "offerRent",
-    label: "Rent a place out",
-    icon: "#ico36_uc_realestate",
-    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
-    draft: {
-      ...emptyDraft,
-      is: {
-        type: "won:RealEstateRentOffer",
-        title: "For Rent",
-        tags: ["RentOutRealEstate"],
-      },
-      seeks: {
-        type: "won:RealEstateRentDemand",
-        tags: ["SearchRealEstateToRent"],
-      },
-    },
-    isDetails: {
-      title: { ...details.title },
-      description: { ...details.description },
-      location: {
-        ...details.location,
-        mandatory: true,
-      },
-      floorSize: {
-        ...realEstateFloorSizeDetail,
-        mandatory: true,
-      },
-      numberOfRooms: {
-        ...realEstateNumberOfRoomsDetail,
-        mandatory: true,
-      },
-      features: { ...realEstateFeaturesDetail },
-      rent: {
-        ...realEstateRentDetail,
-        mandatory: true,
-      },
-    },
-    seeksDetails: undefined,
-  },
-  // searchBuy: {},
-  // offerBuy: {},
-};
-
-const transportUseCases = {
-  transportDemand: {
-    identifier: "transportDemand",
-    label: "Send something",
-    icon: "#ico36_uc_transport_demand",
-    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
-    draft: {
-      ...emptyDraft,
-      is: { title: "Want to send something" },
-    },
-    isDetails: {
-      title: { ...details.title },
-      content: {
-        ...details.description,
-        identifier: "content",
-        label: "Content",
-        placeholder: "Provide information about what should be transported",
-        parseToRDF: function({ value }) {
-          if (!value) {
-            return { "s:name": undefined };
-          } else {
-            return { "@type": "s:Product", "s:name": value };
-          }
-        },
-        parseFromRDF: function(jsonLDImm) {
-          const content = won.parseFrom(jsonLDImm, ["s:name"], "xsd:string");
-          const type = get(jsonLDImm, "@type");
-          if (content && type === "s:Product") {
-            return content;
-          } else {
-            return undefined;
-          }
-        },
-      },
-      weight: {
-        ...abstractDetails.number,
-        identifier: "weight",
-        label: "Weight in kg",
-        icon: "#ico36_detail_weight",
-        parseToRDF: function({ value }) {
-          if (!isValidNumber(value)) {
-            return { "s:weight": undefined };
-          } else {
-            return {
-              "@type": "s:Product",
-              "s:weight": {
-                "@type": "s:QuantitativeValue",
-                "s:value": [{ "@value": value, "@type": "xsd:float" }],
-                "s:unitCode": "KGM",
-              },
-            };
-          }
-        },
-        parseFromRDF: function(jsonLDImm) {
-          const w = won.parseFrom(
-            jsonLDImm,
-            ["s:weight", "s:value"],
-            "xsd:float"
-          );
-          const unit = getInFromJsonLd(
-            jsonLDImm,
-            ["s:weight", "s:unitCode"],
-            won.defaultContext
-          );
-
-          if (!w) {
-            return undefined;
-          } else {
-            if (unit === "KGM") {
-              return w + "kg";
-            } else if (unit === "GRM") {
-              return w + "g";
-            } else if (!unit) {
-              return w + " (no unit specified)";
-            }
-            return w + " " + unit;
-          }
-        },
-        generateHumanReadable: function({ value, includeLabel }) {
-          if (isValidNumber(value)) {
-            return (includeLabel ? this.label + ": " + value : value) + "kg";
-          }
-          return undefined;
-        },
-      },
-      length: {
-        ...abstractDetails.number,
-        identifier: "length",
-        label: "Length in cm",
-        icon: "#ico36_detail_measurement",
-        parseToRDF: function({ value }) {
-          if (!isValidNumber(value)) {
-            return { "s:length": undefined };
-          } else {
-            return {
-              "@type": "s:Product",
-              "s:length": {
-                "@type": "s:QuantitativeValue",
-                "s:value": [{ "@value": value, "@type": "xsd:float" }],
-                "s:unitCode": "CMT",
-              },
-            };
-          }
-        },
-        parseFromRDF: function(jsonLDImm) {
-          const l = won.parseFrom(
-            jsonLDImm,
-            ["s:length", "s:value"],
-            "xsd:float"
-          );
-          const unit = getInFromJsonLd(
-            jsonLDImm,
-            ["s:length", "s:unitCode"],
-            won.defaultContext
-          );
-
-          if (!l) {
-            return undefined;
-          } else {
-            if (unit === "CMT") {
-              return l + "cm";
-            } else if (unit === "MTR") {
-              return l + "m";
-            } else if (!unit) {
-              return l + " (no unit specified)";
-            }
-            return l + " " + unit;
-          }
-        },
-        generateHumanReadable: function({ value, includeLabel }) {
-          if (isValidNumber(value)) {
-            return (includeLabel ? this.label + ": " + value : value) + "cm";
-          }
-          return undefined;
-        },
-      },
-      width: {
-        ...abstractDetails.number,
-        identifier: "width",
-        label: "Width in cm",
-        icon: "#ico36_detail_measurement",
-        parseToRDF: function({ value }) {
-          if (!isValidNumber(value)) {
-            return { "s:width": undefined };
-          } else {
-            return {
-              "@type": "s:Product",
-              "s:width": {
-                "@type": "s:QuantitativeValue",
-                "s:value": [{ "@value": value, "@type": "xsd:float" }],
-                "s:unitCode": "CMT",
-              },
-            };
-          }
-        },
-        parseFromRDF: function(jsonLDImm) {
-          const w = won.parseFrom(
-            jsonLDImm,
-            ["s:width", "s:value"],
-            "xsd:float"
-          );
-          const unit = getInFromJsonLd(
-            jsonLDImm,
-            ["s:width", "s:unitCode"],
-            won.defaultContext
-          );
-
-          if (!w) {
-            return undefined;
-          } else {
-            if (unit === "CMT") {
-              return w + "cm";
-            } else if (unit === "MTR") {
-              return w + "m";
-            } else if (!unit) {
-              return w + " (no unit specified)";
-            }
-            return w + " " + unit;
-          }
-        },
-        generateHumanReadable: function({ value, includeLabel }) {
-          if (isValidNumber(value)) {
-            return (includeLabel ? this.label + ": " + value : value) + "cm";
-          }
-          return undefined;
-        },
-      },
-      height: {
-        ...abstractDetails.number,
-        identifier: "height",
-        label: "Height in cm",
-        icon: "#ico36_detail_measurement",
-        parseToRDF: function({ value }) {
-          if (!isValidNumber(value)) {
-            return { "s:height": undefined };
-          } else {
-            return {
-              "@type": "s:Product",
-              "s:height": {
-                "@type": "s:QuantitativeValue",
-                "s:value": [{ "@value": value, "@type": "xsd:float" }],
-                "s:unitCode": "CMT",
-              },
-            };
-          }
-        },
-        parseFromRDF: function(jsonLDImm) {
-          const h = won.parseFrom(
-            jsonLDImm,
-            ["s:height", "s:value"],
-            "xsd:float"
-          );
-          const unit = getInFromJsonLd(
-            jsonLDImm,
-            ["s:height", "s:unitCode"],
-            won.defaultContext
-          );
-
-          if (!h) {
-            return undefined;
-          } else {
-            if (unit === "CMT") {
-              return h + "cm";
-            } else if (unit === "MTR") {
-              return h + "m";
-            } else if (!unit) {
-              return h + " (no unit specified)";
-            }
-            return h + " " + unit;
-          }
-        },
-        generateHumanReadable: function({ value, includeLabel }) {
-          if (isValidNumber(value)) {
-            return (includeLabel ? this.label + ": " + value : value) + "cm";
-          }
-          return undefined;
-        },
-      },
-      tags: { ...details.tags },
-    },
-    seeksDetails: {
-      travelAction: { ...details.travelAction },
-    },
-  },
-  transportOffer: {
-    identifier: "transportOffer",
-    label: "Offer goods transport",
-    icon: "#ico36_uc_transport_offer",
-    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
-    draft: {
-      ...emptyDraft,
-      is: { title: "Transportation offer" },
-      searchString: "transport", // TODO: replace this with a query
-    },
-    isDetails: {
-      title: { ...details.title },
-      location: { ...details.location },
-    },
-    seeksDetails: {
-      tags: { ...details.tags },
-      description: { ...details.description },
-    },
-    // generateQuery: (draft, resultName) => {
-    //   const filterStrings = [];
-    //   const prefixes = {
-    //     s: won.defaultContext["s"],
-    //     won: won.defaultContext["won"],
-    //   };
-
-    //   let queryTemplate =
-    //     `
-    //     ${prefixesString(prefixes)}
-    //     SELECT DISTINCT ${resultName}
-    //     WHERE {
-    //     ${resultName}
-    //       won:is ?is.
-    //       ${filterStrings && filterStrings.join(" ")}
-    //     }` + (location ? `ORDER BY ASC(?geoDistance)` : "");
-    //   return new SparqlParser().parse(queryTemplate);
-    // },
-  },
-};
-
-const mobilityUseCases = {
-  liftDemand: {
-    identifier: "liftDemand",
-    label: "Need a Lift",
-    icon: "#ico36_uc_route_demand",
-    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
-    draft: {
-      ...emptyDraft,
-      is: { title: "Need a lift", tags: "search-lift" },
-      searchString: "offer-lift",
-    },
-    // TODO: amount of people? other details?
-    isDetails: {
-      title: { ...details.title },
-      description: { ...details.description },
-    },
-    seeksDetails: {
-      fromDatetime: { ...details.fromDatetime },
-      travelAction: { ...details.travelAction },
-    },
-    generateQuery: (draft, resultName) => {
-      const fromLocation = getIn(draft, [
-        "seeks",
-        "travelAction",
-        "fromLocation",
-      ]);
-      const toLocation = getIn(draft, ["seeks", "travelAction", "toLocation"]);
-
-      const baseFilter = {
-        prefixes: {
-          won: won.defaultContext["won"],
-        },
-        operations: [
-          `${resultName} a won:Need.`,
-          `${resultName} won:isInState won:Active. { {${resultName} a <http://dbpedia.org/page/Ridesharing>.  } union { ${resultName} a s:TaxiService} }`,
-          `${resultName} won:is ?is.`,
-        ],
-      };
-
-      const locationFilter = filterInVicinity(
-        "?location",
-        fromLocation,
-        /*radius=*/ 100
-      );
-      const fromLocationFilter = filterInVicinity(
-        "?fromLocation",
-        fromLocation,
-        /*radius=*/ 100
-      );
-      const toLocationFilter = filterInVicinity(
-        "?toLocation",
-        toLocation,
-        /*radius=*/ 100
-      );
-
-      const union = operations => {
-        if (!operations || operations.length === 0) {
-          return "";
-        } else {
-          return "{" + operations.join("} UNION {") + "}";
-        }
-      };
-      const filterAndJoin = (arrayOfStrings, seperator) =>
-        arrayOfStrings.filter(str => str).join(seperator);
-
-      const locationFilters = {
-        prefixes: locationFilter.prefixes,
-        operations: union([
-          filterAndJoin(
-            [
-              fromLocation &&
-                `${resultName} a <http://dbpedia.org/page/Ridesharing>. ?is won:travelAction/s:fromLocation ?fromLocation. `,
-              fromLocation && fromLocationFilter.operations.join(" "),
-              toLocation && "?is won:travelAction/s:toLocation ?toLocation.",
-              toLocation && toLocationFilter.operations.join(" "),
-            ],
-            " "
-          ),
-          filterAndJoin(
-            [
-              location &&
-                `${resultName} a s:TaxiService . ?is won:hasLocation ?location .`,
-              location && locationFilter.operations.join(" "),
-            ],
-            " "
-          ),
-        ]),
-      };
-
-      const concatenatedFilter = concatenateFilters([
-        baseFilter,
-        locationFilters,
-      ]);
-
-      return sparqlQuery({
-        prefixes: concatenatedFilter.prefixes,
-        selectDistinct: resultName,
-        where: concatenatedFilter.operations,
-        orderBy: [
-          {
-            order: "ASC",
-            variable: "?location_geoDistance",
-          },
-        ],
-      });
-    },
-  },
-  taxiOffer: {
-    identifier: "taxiOffer",
-    label: "Offer Taxi Service",
-    icon: "#ico36_uc_taxi_offer",
-    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
-    draft: {
-      ...emptyDraft,
-      is: { title: "Taxi", type: "s:TaxiService" },
-    },
-    isDetails: {
-      title: { ...details.title },
-      description: { ...details.description },
-      location: { ...details.location },
-    },
-    generateQuery: (draft, resultName) => {
-      const location = getIn(draft, ["is", "location"]);
-      const filters = [
-        {
-          // to select seeks-branch
-          prefixes: {
-            won: won.defaultContext["won"],
-          },
-          operations: [
-            `${resultName} a won:Need.`,
-            `${resultName} won:seeks ?seeks.`,
-            location && "?seeks won:travelAction/s:fromLocation ?location.",
-          ],
-        },
-
-        filterInVicinity("?location", location, /*radius=*/ 100),
-      ];
-
-      const concatenatedFilter = concatenateFilters(filters);
-
-      return sparqlQuery({
-        prefixes: concatenatedFilter.prefixes,
-        selectDistinct: resultName,
-        where: concatenatedFilter.operations,
-        orderBy: [
-          {
-            order: "ASC",
-            variable: "?location_geoDistance",
-          },
-        ],
-      });
-    },
-  },
-  rideShareOffer: {
-    identifier: "rideShareOffer",
-    label: "Offer to Share a Ride",
-    icon: "#ico36_uc_taxi_offer",
-    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
-    draft: {
-      ...emptyDraft,
-      is: {
-        title: "Share a Ride",
-        type: "http://dbpedia.org/page/Ridesharing",
-      },
-    },
-    isDetails: {
-      title: { ...details.title },
-      description: { ...details.description },
-      fromDatetime: { ...details.fromDatetime },
-      throughDatetime: { ...details.throughDatetime },
-      travelAction: { ...details.travelAction },
-    },
-    generateQuery: (draft, resultName) => {
-      const toLocation = getIn(draft, ["is", "travelAction", "toLocation"]);
-      const fromLocation = getIn(draft, ["is", "travelAction", "fromLocation"]);
-
-      const fromTime = getIn(draft, ["is", "fromDatetime"]);
-      const filters = [
-        {
-          // to select seeks-branch
-          prefixes: {
-            won: won.defaultContext["won"],
-          },
-          operations: [
-            `${resultName} a won:Need.`,
-            `${resultName} won:seeks ?seeks.`,
-            fromLocation &&
-              "?seeks won:travelAction/s:fromLocation ?fromLocation.",
-            toLocation && "?seeks won:travelAction/s:toLocation ?toLocation.",
-            isValidDate(fromTime) && "?seeks s:validFrom ?starttime",
-          ],
-        },
-
-        filterInVicinity("?fromLocation", fromLocation, /*radius=*/ 100),
-
-        filterInVicinity("?toLocation", toLocation, /*radius=*/ 100),
-
-        filterAboutTime("?starttime", fromTime, 12 /* hours before and after*/),
-      ];
-
-      const concatenatedFilter = concatenateFilters(filters);
-
-      return sparqlQuery({
-        prefixes: concatenatedFilter.prefixes,
-        selectDistinct: resultName,
-        where: concatenatedFilter.operations,
-        orderBy: [
-          {
-            order: "ASC",
-            variable: "?fromLocation_geoDistance",
-          },
-        ],
-      });
-    },
-  },
-};
-
-/**
- * band musician use cases
- */
-const musicianUseCases = {
-  findBand: {
-    identifier: "findBand",
-    label: "Find Band",
-    icon: "#ico36_uc_find_band",
-    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
-    draft: {
-      ...emptyDraft,
-      is: {
-        title: "I'm looking for a band!",
-        //tags: ["musician", "band"],
-      },
-      searchString: "band",
-    },
-    isDetails: {
-      title: { ...details.title },
-      description: { ...details.description },
-      instruments: {
-        ...instrumentsDetail,
-        //mandatory: true,
-      },
-    },
-    seeksDetails: {
-      description: { ...details.description },
-      genres: { ...genresDetail },
-      location: { ...details.location },
-    },
-  },
-  findMusician: {
-    identifier: "findMusician",
-    label: "Find Musician",
-    icon: "#ico36_uc_find_musician",
-    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
-    draft: {
-      ...emptyDraft,
-      is: {
-        title: "Looking for a Musician!",
-        //tags: ["band", "musician"],
-      },
-      searchString: "musician",
-    },
-    isDetails: {
-      title: { ...details.title },
-      description: { ...details.description },
-      genres: { ...genresDetail },
-      location: { ...details.location },
-    },
-    seeksDetails: {
-      description: { ...details.description },
-      instruments: {
-        ...instrumentsDetail,
-        //mandatory: true,
-      },
-    },
-  },
-  findRehearsalRoom: {
-    identifier: "findRehearsalRoom",
-    label: "Find Rehearsal Room",
-    icon: "#ico36_uc_realestate",
-    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
-    draft: {
-      ...emptyDraft,
-      is: {
-        title: "Looking for Rehearsal Room!",
-        tags: ["SearchRehearsal"],
-      },
-      seeks: {
-        tags: ["OfferRehearsal"],
-      },
-      searchString: "Rehearsal Room",
-    },
-    isDetails: undefined,
-    seeksDetails: {
-      location: { ...details.location },
-      floorSizeRange: { ...realEstateFloorSizeRangeDetail },
-      features: {
-        ...realEstateFeaturesDetail,
-        placeholder: "e.g. PA, Drumkit",
-      },
-      rentRange: { ...realEstateRentRangeDetail },
-      fromDatetime: { ...details.fromDatetime },
-      throughDatetime: { ...details.throughDatetime },
-    },
-  },
-  offerRehearsalRoom: {
-    identifier: "OfferRehearsalRoom",
-    label: "Offer Rehearsal Room",
-    icon: "#ico36_uc_realestate",
-    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
-    draft: {
-      ...emptyDraft,
-      is: {
-        title: "Offer Rehearsal Room!",
-        tags: ["OfferRehearsal"],
-      },
-      seeks: {
-        tags: ["SearchRehearsal"],
-      },
-    },
-    isDetails: {
-      title: { ...details.title },
-      description: { ...details.description },
-      location: {
-        ...details.location,
-        mandatory: true,
-      },
-      floorSize: {
-        ...realEstateFloorSizeDetail,
-        mandatory: true,
-      },
-      numberOfRooms: {
-        ...realEstateNumberOfRoomsDetail,
-        mandatory: true,
-      },
-      features: {
-        ...realEstateFeaturesDetail,
-        placeholder: "e.g. PA, Drumkit",
-      },
-      rent: {
-        ...realEstateRentDetail,
-        mandatory: true,
-      },
-      fromDatetime: { ...details.fromDatetime },
-      throughDatetime: { ...details.throughDatetime },
-    },
-
-    seeksDetails: undefined,
   },
 };
