+++ webofneeds/won-owner-webapp/src/main/webapp/config/usecases/professional.js
@@ -1,0 +1,183 @@
+/**
+ * Created by fsuda on 18.09.2018.
+ */
+import { details, emptyDraft } from "detailDefinitions";
+import { interestsDetail, skillsDetail } from "personDetails";
+import { findLatestIntervallEndInJsonLdOrNowAndAddMillis } from "../../app/won-utils.js";
+
+export const professionalUseCases = {
+  getToKnow: {
+    identifier: "getToKnow",
+    label: "Find people",
+    icon: "#ico36_uc_find_people",
+    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
+    draft: {
+      ...emptyDraft,
+      is: {
+        title: "I'm up for meeting new people!",
+        tags: ["meetup"],
+      },
+      searchString: "meetup",
+    },
+    isDetails: {
+      title: { ...details.title },
+      description: { ...details.description },
+      location: { ...details.location },
+      person: { ...details.person },
+      skills: { ...skillsDetail },
+      interests: { ...interestsDetail },
+    },
+    seeksDetails: {
+      description: { ...details.description },
+      location: { ...details.location },
+      skills: { ...skillsDetail },
+      interests: { ...interestsDetail },
+    },
+  },
+  phdIs: {
+    identifier: "phdIs",
+    label: "Offer a PhD position",
+    icon: "#ico36_uc_phd",
+    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
+    draft: {
+      ...emptyDraft,
+      is: {
+        title: "I'm offering a PhD position!",
+        tags: ["offer-phd"],
+      },
+      searchString: "search-phd",
+    },
+    isDetails: {
+      title: { ...details.title },
+      description: { ...details.description },
+      location: { ...details.location },
+    },
+    seeksDetails: {
+      skills: { ...skillsDetail, placeholder: "" }, // TODO: find good placeholders
+      interests: { ...interestsDetail, placeholder: "" },
+    },
+  },
+  phdSeeks: {
+    identifier: "phdSeeks",
+    label: "Find a PhD position",
+    icon: "#ico36_uc_phd",
+    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
+    draft: {
+      ...emptyDraft,
+      is: {
+        title: "I'm looking for a PhD position!",
+        tags: ["search-phd"],
+      },
+      searchString: "offer-phd",
+    },
+    isDetails: {
+      title: { ...details.title },
+      description: { ...details.description },
+      person: { ...details.person },
+      skills: { ...skillsDetail, placeholder: "" }, // TODO: find good placeholders
+      interests: { ...interestsDetail, placeholder: "" },
+    },
+    seeksDetails: {
+      description: { ...details.description },
+      location: { ...details.location },
+    },
+  },
+  postDocIs: {
+    identifier: "postDocIs",
+    label: "Offer a PostDoc position",
+    icon: "#ico36_uc_postdoc",
+    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
+    draft: {
+      ...emptyDraft,
+      is: {
+        title: "I'm offering a PostDoc position!",
+        tags: ["offer-postdoc"],
+      },
+      searchString: "search-postdoc",
+    },
+    isDetails: {
+      title: { ...details.title },
+      description: { ...details.description },
+      location: { ...details.location },
+    },
+    seeksDetails: {
+      skills: { ...skillsDetail, placeholder: "" }, // TODO: find good placeholders
+      interests: { ...interestsDetail, placeholder: "" },
+    },
+  },
+  postDocSeeks: {
+    identifier: "postDocSeeks",
+    label: "Find a PostDoc position",
+    icon: "#ico36_uc_postdoc",
+    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
+    draft: {
+      ...emptyDraft,
+      is: {
+        title: "I'm looking for a PostDoc position!",
+        tags: ["search-postdoc"],
+      },
+      searchString: "offer-postdoc",
+    },
+    isDetails: {
+      title: { ...details.title },
+      description: { ...details.description },
+      person: { ...details.person },
+      skills: { ...skillsDetail, placeholder: "" }, // TODO: find good placeholders
+      interests: { ...interestsDetail, placeholder: "" },
+    },
+    seeksDetails: {
+      description: { ...details.description },
+      location: { ...details.location },
+    },
+  },
+  consortiumIs: {
+    identifier: "consortiumIs",
+    label: "Offer slot in a project consortium",
+    icon: "#ico36_uc_consortium-offer",
+    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
+    draft: {
+      ...emptyDraft,
+      is: {
+        title: "Offering a slot in a project consortium",
+        tags: ["offer-consortium"],
+      },
+      searchString: "search-consortium",
+    },
+    isDetails: {
+      title: { ...details.title },
+      description: { ...details.description },
+      location: { ...details.location },
+    },
+    seeksDetails: {
+      description: { ...details.description },
+      location: { ...details.location },
+      skills: { ...skillsDetail, placeholder: "" }, // TODO: find good placeholders
+      interests: { ...interestsDetail, placeholder: "" },
+    },
+  },
+  consortiumSeeks: {
+    identifier: "consortiumSeeks",
+    label: "Find a project consortium to join",
+    icon: "#ico36_uc_consortium-search",
+    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
+    draft: {
+      ...emptyDraft,
+      is: {
+        title: "Looking for a slot in a project consortium",
+        tags: ["search-consortium"],
+      },
+      searchString: "offer-consortium",
+    },
+    isDetails: {
+      title: { ...details.title },
+      description: { ...details.description },
+      location: { ...details.location },
+      skills: { ...skillsDetail, placeholder: "" }, // TODO: find good placeholders
+      interests: { ...interestsDetail, placeholder: "" },
+    },
+    seeksDetails: {
+      description: { ...details.description },
+      location: { ...details.location },
+    },
+  },
+};
