+++ webofneeds/won-owner-webapp/src/main/webapp/config/usecases/social.js
@@ -1,0 +1,97 @@
+/**
+ * Created by fsuda on 18.09.2018.
+ */
+import { details, emptyDraft } from "detailDefinitions";
+import { findLatestIntervallEndInJsonLdOrNowAndAddMillis } from "../../app/won-utils.js";
+import { interestsDetail } from "personDetails";
+
+export const socialUseCases = {
+  breakfast: {
+    identifier: "breakfast",
+    label: "Get breakfast",
+    icon: "#ico36_uc_breakfast",
+    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
+    draft: {
+      ...emptyDraft,
+      is: {
+        title: "I'm up for breakfast! Any plans?",
+        tags: ["breakfast"],
+      },
+      seeks: { title: "breakfast" },
+      searchString: "breakfast",
+    },
+    isDetails: {
+      title: { ...details.title },
+      description: { ...details.description },
+      fromDatetime: { ...details.fromDatetime },
+      throughDatetime: { ...details.throughDatetime },
+      location: { ...details.location },
+      interests: { ...interestsDetail },
+    },
+    seeksDetails: undefined,
+  },
+  lunch: {
+    identifier: "lunch",
+    label: "Get lunch",
+    icon: "#ico36_uc_meal-half",
+    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
+    draft: {
+      ...emptyDraft,
+      is: {
+        title: "I'm up for lunch! Any plans?",
+        tags: ["lunch"],
+      },
+      searchString: "lunch",
+    },
+    isDetails: {
+      title: { ...details.title },
+      description: { ...details.description },
+      fromDatetime: { ...details.fromDatetime },
+      throughDatetime: { ...details.throughDatetime },
+      location: { ...details.location },
+      interests: { ...interestsDetail },
+    },
+    seeksDetails: undefined,
+  },
+  afterparty: {
+    identifier: "afterparty",
+    label: "Go out",
+    icon: "#ico36_uc_drinks",
+    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
+    draft: {
+      ...emptyDraft,
+      is: {
+        title: "I'm up for partying! Any plans?",
+        tags: ["afterparty"],
+      },
+      searchString: "afterparty",
+    },
+    isDetails: {
+      title: { ...details.title },
+      fromDatetime: { ...details.fromDatetime },
+      throughDatetime: { ...details.throughDatetime },
+      description: { ...details.description },
+      location: { ...details.location },
+      interests: { ...interestsDetail },
+    },
+  },
+  sightseeing: {
+    identifier: "sightseeing",
+    label: "Go sightseeing",
+    icon: "#ico36_uc_sightseeing",
+    doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
+    draft: {
+      ...emptyDraft,
+      is: { tags: ["sightseeing"] },
+      searchString: "sightseeing",
+    },
+    isDetails: {
+      title: { ...details.title },
+      description: { ...details.description },
+      fromDatetime: { ...details.fromDatetime },
+      throughDatetime: { ...details.throughDatetime },
+      location: { ...details.location },
+      interests: { ...interestsDetail },
+    },
+  },
+};
